Move auth folder to pwa Fixes #11
