+++ apps/holder/projects/browser-extension/src/app/app.config.ts
@@ -7,6 +7,8 @@ import { MAT_DIALOG_DEFAULT_OPTIONS } from '@angular/material/dialog';
 import { provideOAuthClient } from 'angular-oauth2-oidc';
 import { environment } from '../environments/environment';
 import { ApiModule, Configuration } from '../../../shared/api/kms';
+import { AuthServiceInterface } from '../../../shared/settings/settings.component';
+import { AuthService } from './auth/auth.service';
 
 function getConfiguration() {
   return new Configuration({
@@ -27,6 +29,10 @@ export const appConfig: ApplicationConfig = {
     provideOAuthClient(),
     importProvidersFrom(ApiModule, HttpClientModule),
     { provide: MAT_DIALOG_DEFAULT_OPTIONS, useValue: { hasBackdrop: true } },
+    {
+      provide: AuthServiceInterface,
+      useClass: AuthService,
+    },
     {
       provide: Configuration,
       useFactory: getConfiguration,
